% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download dataset with ID}
\usage{
download_data(name)
}
\arguments{
\item{name}{ID of dataset}
}
\value{
an object containing relevant dataset files
}
\description{
Download dataset with ID
}
\examples{

download_data(name = 'dd-monty-python-2021-03-2')

}
