% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_categories.R
\name{get_categories}
\alias{get_categories}
\title{Get All Categories}
\usage{
get_categories(within = "categories", about = FALSE)
}
\arguments{
\item{within}{Main category to search. Defaults to `categories`. Arguments can include `categories`, `cities`, `states`.}

\item{about}{Sub-category to search within the main category. Can include categories like `economics`, `health`, city names like `delhi` or names of states like `assam`.}
}
\value{
a character vector of items.
}
\description{
Get All Categories
}
\examples{

get_categories()

get_categories(within = "states")

get_categories(within = "states", about = "assam")

}
