% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.r
\name{get_datasets}
\alias{get_datasets}
\title{Get datasets within categories}
\usage{
get_datasets(name, within = FALSE, about = FALSE)
}
\arguments{
\item{name}{If you know the ID of the dataset you're looking for, use the name parameter}

\item{within}{The main category to look within. Defaults to `Categories`. Can include `States` or `Cities` too.}

\item{about}{Specify your topic. Can include terms like `Entertainment`, `Sports`, `assam`}
}
\value{
a tibble of the search results
}
\description{
Get datasets within categories
}
\examples{
get_datasets()

get_datasets(within = 'Categories')

get_datasets(within = 'States', about = 'assam')

}
