% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a Second}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character vector}

\item{split}{what to split by}
}
\value{
A character vector
}
\description{
Split a Second
}
\examples{
 x <- "alfa,bravo,charlie,delta"
 strsplit(x, split = ",")
}
